refactor(harness): extract temporary resource discovery owner
